+++ src/webgrapher.webui/CHANGELOG.md
@@ -4,9 +4,9 @@ The following tools were used to generate this project:
 - create-vite
 
 The following steps were used to generate this project:
-- Create vue project with create-vite: `npm init --yes vue@latest webmapper.webui -- --eslint `.
+- Create vue project with create-vite: `npm init --yes vue@latest webgrapher.webui -- --eslint `.
 - Updating `vite.config.js` with port.
-- Create project file (`webmapper.webui.esproj`).
+- Create project file (`webgrapher.webui.esproj`).
 - Create `launch.json` to enable debugging.
 - Add project to solution.
 - Write this file.